refactor: Merge `NormalHeader` and `Header`
