+++ source/components/Header/index.tsx
@@ -1,4 +1,3 @@
 export * from './Header';
 export { AccountHeader } from './AccountHeader';
-export { NormalHeader } from './NormalHeader';
 export * from './LogoHeader';
